test(archive/tar): isolate unix testable-only tests.
