+++ compression/archive/tar/extract_test.go
@@ -11,9 +11,7 @@ import (
 	"fmt"
 	"io/fs"
 	"os"
-	"runtime"
 	"strings"
-	"syscall"
 	"testing"
 
 	"github.com/stretchr/testify/require"
@@ -422,116 +420,6 @@ func TestExtract_WithRestoreTimes(t *testing.T) {
 	fi, err := root.Lstat("file.txt")
 	require.NoError(t, err)
 	require.NotZero(t, fi.ModTime())
-}
-
-func TestExtract_WithRestoreOwner(t *testing.T) {
-	t.Parallel()
-
-	// Skip on Windows
-	if runtime.GOOS == "windows" {
-		t.Skip("skipping test on windows")
-	}
-
-	out := &bytes.Buffer{}
-	b, err := builder.New(out).With(
-		builder.File("file.txt", strings.NewReader("hello, world"),
-			builder.WithUID(os.Getuid()),
-			builder.WithGID(os.Getgid()),
-		),
-	)
-	require.NoError(t, err)
-	require.NotNil(t, b)
-	require.NoError(t, b.Close())
-
-	tmpDir := t.TempDir()
-	require.NoError(t, Extract(out, tmpDir, WithRestoreOwner(true)))
-
-	// Create a new file system
-	root, err := vfs.Chroot(tmpDir)
-	require.NoError(t, err)
-
-	// Check the file
-	fi, err := root.Lstat("file.txt")
-	require.NoError(t, err)
-	require.NotNil(t, fi.Sys())
-	require.IsType(t, &syscall.Stat_t{}, fi.Sys())
-	require.Equal(t, os.Getuid(), int(fi.Sys().(*syscall.Stat_t).Uid))
-	require.Equal(t, os.Getgid(), int(fi.Sys().(*syscall.Stat_t).Gid))
-}
-
-func TestExtract_WithRestoreOwner_WithRemapper(t *testing.T) {
-	t.Parallel()
-
-	// Skip on Windows
-	if runtime.GOOS == "windows" {
-		t.Skip("skipping test on windows")
-	}
-
-	t.Run("remap", func(t *testing.T) {
-		out := &bytes.Buffer{}
-		b, err := builder.New(out).With(
-			builder.File("file.txt", strings.NewReader("hello, world"),
-				builder.WithUID(0),
-				builder.WithGID(0),
-			),
-		)
-		require.NoError(t, err)
-		require.NotNil(t, b)
-		require.NoError(t, b.Close())
-
-		tmpDir := t.TempDir()
-		require.NoError(t, Extract(out, tmpDir, WithRestoreOwner(true), WithUIDGIDMapperFunc(func(uid, gid int) (int, int, error) {
-			return os.Getuid(), os.Getgid(), nil
-		})))
-
-		// Create a new file system
-		root, err := vfs.Chroot(tmpDir)
-		require.NoError(t, err)
-
-		// Check the file
-		fi, err := root.Lstat("file.txt")
-		require.NoError(t, err)
-		require.NotNil(t, fi.Sys())
-		require.IsType(t, &syscall.Stat_t{}, fi.Sys())
-		require.Equal(t, os.Getuid(), int(fi.Sys().(*syscall.Stat_t).Uid))
-		require.Equal(t, os.Getgid(), int(fi.Sys().(*syscall.Stat_t).Gid))
-	})
-
-	t.Run("remap_error", func(t *testing.T) {
-		out := &bytes.Buffer{}
-		b, err := builder.New(out).With(
-			builder.File("file.txt", strings.NewReader("hello, world"),
-				builder.WithUID(0),
-				builder.WithGID(0),
-			),
-		)
-		require.NoError(t, err)
-		require.NotNil(t, b)
-		require.NoError(t, b.Close())
-
-		tmpDir := t.TempDir()
-		require.Error(t, Extract(out, tmpDir, WithRestoreOwner(true), WithUIDGIDMapperFunc(func(uid, gid int) (int, int, error) {
-			return -1, -1, fmt.Errorf("unable to map UID/GID")
-		})))
-	})
-
-	t.Run("remap_invalid", func(t *testing.T) {
-		out := &bytes.Buffer{}
-		b, err := builder.New(out).With(
-			builder.File("file.txt", strings.NewReader("hello, world"),
-				builder.WithUID(0),
-				builder.WithGID(0),
-			),
-		)
-		require.NoError(t, err)
-		require.NotNil(t, b)
-		require.NoError(t, b.Close())
-
-		tmpDir := t.TempDir()
-		require.Error(t, Extract(out, tmpDir, WithRestoreOwner(true), WithUIDGIDMapperFunc(func(uid, gid int) (int, int, error) {
-			return -1, -1, nil
-		})))
-	})
 }
 
 func TestExtract_WithRestoreMode(t *testing.T) {
